Added scriptable objects for combo creation. Only two test combos are created. Ui will reflect the current combo image, name, and input sequence. Currently no way to cycle through combos just yet.
Next steps are figuring out how to do more input logic. I need more visual feedback on an incorrect and correct input. with the sequence resetting on an incorrect one. I need to create all of the combos that will exists in the game  so I can start working with rounds and how to progress those rounds, as well as changing the order of the combos in each round, and possibly adding more difficult ones.