Created a game manager to keep track of the timer, score, and round. Rounds now generate random combos from a list of all the combos. I created 8 total combos for now to test with. There are some known bugs since its not done. Timer need fixing and turned into a bar instead of a text element. Id like a screen to take a moment between rounds to show the score.
I've noticed a Unity bug that happens when I select a gameObject in the Hierarchy. The bug is: NullReferenceException: SerializedObject of SerializedProperty has been Disposed.
After some research it appears this bug has to do with the version Im using and is fixed in older versions. Hopefiully this doesnt cause to much trouble since I have no way to fix it.